% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.survey_mde}
\alias{coef.survey_mde}
\title{Extract Model Coefficients}
\usage{
\method{coef}{survey_mde}(object, ...)
}
\arguments{
\item{object}{an MDE object fitted to survey data by \code{\link[=survey_mpde]{survey_mpde()}} or \code{\link[=survey_mhde]{survey_mhde()}}.}

\item{...}{currently unused.}
}
\description{
Extract coefficients from a fitted MDE.
}
