% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-helper.R
\name{.extract_survey_values}
\alias{.extract_survey_values}
\title{Extract the Numeric Values from a Survey Design}
\usage{
.extract_survey_values(x, design, na.rm = FALSE)
}
\description{
Extract the Numeric Values from a Survey Design
}
\keyword{internal}
