% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{.kde}
\alias{.kde}
\title{Evaluate the Univariate Weighted KDE}
\usage{
.kde(x, wgts, evalpts, bw, kernel)
}
\arguments{
\item{x}{observations (numeric vector)}

\item{wgts}{weights (numeric vector of same length as \code{x})}

\item{evalpts}{\strong{ordered(!)} evaluations points (numeric vector)}

\item{bw}{bandwidth (positive scalar)}

\item{kernel}{the first letter of the kernel function to use:
e ... Epanechnikov,
b ... Biweight
t ... Triangular
r ... Rectangular}
}
\description{
None of the parameters are checked! They are assumed to be of the right type and size.
}
\keyword{internal}
