% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov-internal.R
\name{.vcov}
\alias{.vcov}
\title{Calculate Variance-Covariance Matrix for Minimum Divergence Estimators}
\usage{
.vcov(object, type, n, enforce_type = FALSE)
}
\arguments{
\item{object}{a MDE fitted to survey data}

\item{type}{which type of covariance matrix to compute.}

\item{n}{either how to estimate the effective sample size, or (if numeric) the
sample size.}

\item{enforce_type}{should the \code{type} be enforced, or use the model-based estimate
as fallback with a warning.}
}
\description{
Calculate Variance-Covariance Matrix for Minimum Divergence Estimators
}
\keyword{internal}
