% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{epanechnikov_kernel}
\alias{epanechnikov_kernel}
\title{Epanechnikov Kernel for KDE}
\usage{
epanechnikov_kernel(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
the evaluated kernel
}
\description{
Evaluate the Epanechnikov kernel
\deqn{
  k(x) = \frac{3}{4} (1 - x^2) 1_{[|x| \leq 1]}
}
}
