% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-divergence.R
\name{generalized_divergence}
\alias{generalized_divergence}
\title{Define a Generalized Divergence}
\usage{
generalized_divergence(alpha, lambda)
}
\arguments{
\item{alpha, lambda}{the parameters for the generalized divergence.}
}
\value{
an object of type \link{GeneralizedDivergence} or \link{PhiDivergence}, depending
on the combination of parameters \code{alpha} and \code{lambda}.
}
\description{
The family of generalized divergences is defined as
\deqn{
D_{(\alpha,\lambda)}(\hat f \| g_\theta) =
    \frac{1}{1 + \lambda(1 + \alpha)} \int g_\theta^{1 + \alpha}
  - \frac{1+\alpha}{( + \lambda(1 + \alpha)) (\alpha - \lambda(1 - \alpha))} \int g_\theta^{\alpha - \lambda(1 - \alpha)} \hat f^{1 + \lambda(1 + \alpha)}
  + \frac{1}{\alpha - \lambda(1 - \alpha)} \int \hat f^{1 + \alpha}
}
}
