% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-divergence.R
\name{GeneralizedDivergence}
\alias{GeneralizedDivergence}
\title{GeneralizedDivergence Class}
\description{
A class to encapsulate generalized divergences defined as
\deqn{
D_{\alpha,\lambda}(\hat f \| g_\theta) =
    \frac{1}{1 + \lambda(1 + \alpha)} \int \hat f^{1 + \alpha}
  - \frac{1+\alpha}{( + \lambda(1 + \alpha)) (\alpha - \lambda(1 - \alpha))} \int \hat f^{\alpha - \lambda(1 - \alpha)} g_\theta^{1 + \lambda(1 + \alpha)}
  + \frac{1}{\alpha - \lambda(1 - \alpha)} \int g_\theta^{1 + \alpha}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alpha}}{parameter of the divergence generator}

\item{\code{lambda}}{parameter of the divergence generator}

\item{\code{const_A}}{pre-computed \code{1 + lambda * (1 - alpha)}}

\item{\code{const_B}}{pre-computed \code{alpha - lambda * (1 - alpha)}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeneralizedDivergence-new}{\code{GeneralizedDivergence$new()}}
\item \href{#method-GeneralizedDivergence-clone}{\code{GeneralizedDivergence$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneralizedDivergence-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeneralizedDivergence-new}{}}}
\subsection{Method \code{new()}}{
Define a new generalized divergence.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneralizedDivergence$new(alpha, lambda)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{the \eqn{\alpha} parameter of the divergence (between 0 and 1)}

\item{\code{lambda}}{the \eqn{\lambda} parameter of the divergence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneralizedDivergence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeneralizedDivergence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneralizedDivergence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
