% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.survey_mde}
\alias{coef.survey_mde}
\alias{coef.survey_reg_mde}
\alias{sigma.survey_reg_mde}
\alias{vcov}
\alias{vcov.survey_mde}
\title{Extract Information from Fitted Minimum Divergence Estimators}
\usage{
\method{coef}{survey_mde}(object, ...)

\method{coef}{survey_reg_mde}(object, ...)

\method{sigma}{survey_reg_mde}(object, ...)

\method{vcov}{survey_mde}(object, type = c("sandwich", "model"), n = c("score", "kish"), ...)
}
\arguments{
\item{object}{an MDE object fitted to survey data}

\item{...}{currently unused.}

\item{type}{which type of covariance matrix to compute.}

\item{n}{either how to estimate the effective sample size, or (if numeric) the
sample size.}
}
\description{
Methods to extract coefficient estimates and variance/covariance components from
minimum divergence estimators fitted to complex survey data.
}
\details{
\code{coef()} extracts the parameter estimates from the fitted MDE and the coefficient estimates
from the regression MDE.

\code{sigma()} extracts the standard deviation from the fitted regression model.

\code{vcov()} calculates the variance-covariance matrix for fitted MDEs.
}
