% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss-kronrod-quadrature.R
\name{hd_gauss_quadrature}
\alias{hd_gauss_quadrature}
\alias{partition_domain}
\title{Functor to Create Fixed-grid Gauss-Kronrod Quadrature For Hellinger Affinity}
\usage{
hd_gauss_quadrature(
  x,
  wgts,
  bandwidth,
  n_subdivisions = 256,
  poly_order = 20,
  range,
  kernel
)

partition_domain(x, bw, subdivisions, range)
}
\arguments{
\item{x}{numeric vector}

\item{range}{the allowable range of \code{x}. If missing, the assumed allowable range is
\eqn{[\min_i x_i - bw, \max_i x_i + bw]}.}

\item{bw}{bandwidth of the bounded kernel function}

\item{subdivisions}{number of partitions}
}
\value{
a function to evaluate the integral
\deqn{2 - 2 \int \sqrt{ \hat f(x) g(x) }\;\mathrm{d}x}
}
\description{
This creates a function to compute
\deqn{\int \sqrt{ \hat f(x) g(x) }\;\mathrm{d}x,}
or (if \code{log=TRUE})
\deqn{\int \exp\{\frac{1}{2} (\log(\hat f(x)) + \log(g(x)) )\}\;\mathrm{d}x,}
for arbitrary non-negative functions \eqn{g()}.
The function is set up such that \eqn{\hat f} is evaluated exactly 21 times in each
subdivision, and does not need to be re-evaluated for different \eqn{g()}.
}
\details{
\subsection{Partitioning}{

The integration considers only the "relevant" domain spanned by \code{x}.
The span of \code{x} is partitioned into a fixed number of partitions such that the
KDE \eqn{\hat{f}(x) > 0} within each of those partitions and \eqn{\hat{f}(x) = 0} outside
those partitions.
The algorithm for \code{partition_domain()} tries to have all partitions roughly the same length, but the
apportionment is greedy and hence possibly not the globally best solution.
}
}
\keyword{internal}
