% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss-kronrod-basis.R
\name{hd_gauss_quadrature}
\alias{hd_gauss_quadrature}
\title{Functor to Create Fixed-grid Gauss-Kronrod Quadrature For Hellinger Affinity}
\usage{
hd_gauss_quadrature(
  x,
  wgts,
  bandwidth,
  n_subdivisions = 256,
  range,
  kernel = epanechnikov_kernel
)
}
\value{
a function to evaluate the integral
  \deqn{2 - 2 \int \sqrt{ \hat f(x) g(x) }\;\mathrm{d}x}
}
\description{
This creates a function to compute
\deqn{\int \sqrt{ \hat f(x) g(x) }\;\mathrm{d}x}
or
\deqn{\int \exp\{\frac{1}{2} (\log(\hat f(x)) + \log(g(x)) )\}\;\mathrm{d}x}
for arbitrary non-negative functions \eqn{g()}.
The function is set up such that \eqn{\hat f} is evaluated exactly 21 times in each
subdivision, and does not need to be re-evaluated for different \eqn{g()}.
}
\details{
The subdivisions are chosen to cover only regions where \eqn{f(x) > 0} using
a greedy partitioning algorithm.
}
\keyword{internal}
