% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-functions.R
\name{link}
\alias{link}
\alias{Link}
\title{Link Functions}
\usage{
link(mean, ...)
}
\arguments{
\item{mean}{either a character naming one of the built-in link functions (see Details)
or an instance of the \link{Link} R6 class.}

\item{...}{named link functions (either character or an instance of the \link{Link} class),
one for each nuisance parameter..}
}
\description{
The \code{link()} function can be used to define/access link functions linking the location
and nuisance (variance) components to the predictors.

The \code{Link} class encapsulates the link function and it's inverse & derivative
information.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{name/description of the link function}

\item{\code{fun}}{link function mapping the mean (location) parameter of a model family
to the predictors}

\item{\code{inv}}{inverse link function mapping the mean (location) parameter of a model family
to the predictors}

\item{\code{inv_d1}}{first derivative of the inverse link function}

\item{\code{inv_d2}}{second derivative of the inverse link function}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Link-new}{\code{Link$new()}}
\item \href{#method-Link-print}{\code{Link$print()}}
\item \href{#method-Link-clone}{\code{Link$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Link-new"></a>}}
\if{latex}{\out{\hypertarget{method-Link-new}{}}}
\subsection{Method \code{new()}}{
Define a new link function \eqn{\eta} and its inverse \eqn{\eta^{-1}}
to map the location parameter of a model family to the predictors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Link$new(name, fun, inv, inv_d1, inv_d2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the link function}

\item{\code{fun}}{the link function \eqn{\eta(\mu) = f(x)}}

\item{\code{inv}}{the inverse link function \eqn{\mu = \eta^{-1}(f(x))}}

\item{\code{inv_d1}}{the first derivative of the inverse link function,
\eqn{\frac{\partial}{\partial u} \eta^{-1}(u)}}

\item{\code{inv_d2}}{the second derivative of the inverse link function,
\eqn{\frac{\partial^2}{\partial u^2} \eta^{-1}(u)}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Link-print"></a>}}
\if{latex}{\out{\hypertarget{method-Link-print}{}}}
\subsection{Method \code{print()}}{
Print the name and parameters of a model family
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Link$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Link-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Link-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Link$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
