% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhde_gamma.R
\name{mhd_gamma}
\alias{mhd_gamma}
\title{Minimum Hellinger Distance Estimator for the Gamma Model}
\usage{
mhd_gamma(
  x,
  wgts = NULL,
  initial,
  cov_type = c("sandwich", "model", "none"),
  optim_method = "Nelder-Mead",
  optim_control = list()
)
}
\arguments{
\item{x}{univariate observations from the finite population}

\item{initial}{an initial estimate for the shape and scale parameters. If missing,
use the MLE.}

\item{cov_type}{whether to use the sandwich estimator for the covariance matrix
the inverse Fisher information under the model, or not estimate the covariance matrix.}

\item{optim_method, optim_control}{method and control options passed on to [stats::optim()].}

\item{wgt}{sampling weights}
}
\description{
Minimum Hellinger Distance Estimator for the Gamma Model
}
\seealso{
Other Minimum Hellinger Distance Estimator: 
\code{\link{mhd_lognorm}()}
}
\concept{Minimum Hellinger Distance Estimator}
