% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhde_lognormal.R
\name{mhd_lognorm}
\alias{mhd_lognorm}
\title{Minimum Hellinger Distance Estimator for the Lognormal Model}
\usage{
mhd_lognorm(
  x,
  wgts = NULL,
  initial,
  log_transform = FALSE,
  integration_subdivisions = 256,
  optim_method = "Nelder-Mead",
  optim_control = list()
)
}
\arguments{
\item{x}{univariate observations from the finite population}

\item{wgts}{sampling weights}

\item{initial}{an initial estimate for the shape and scale parameters. If missing,
use the MLE.}

\item{log_transform}{whether to log-transform the data before estimation or not.}

\item{integration_subdivisions}{number of partitions to divide the domain of \eqn{\hat f()}
for Gauss-Kronrod quadrature.}

\item{optim_method, optim_control}{method and control options passed on to \code{\link[stats:optim]{stats::optim()}}.}
}
\description{
Compute the MHDE for the parameters \eqn{\mu,\sigma} of the Lognormal model.
}
\seealso{
\code{\link[stats:Lognormal]{stats::dlnorm()}} for the parametrization by \code{meanlog} \eqn{\mu} and
\code{sdlog} \eqn{\sigma}.

Other Minimum Hellinger Distance Estimator: 
\code{\link{mhd_gamma}()}
}
\concept{Minimum Hellinger Distance Estimator}
