% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhde_weibull.R
\name{mhd_weibull}
\alias{mhd_weibull}
\title{Minimum Hellinger Distance Estimator for the Weibull Model}
\usage{
mhd_weibull(
  x,
  wgts = NULL,
  initial,
  integration_subdivisions = 256,
  bw,
  optim_method = "Nelder-Mead",
  optim_control = list()
)
}
\arguments{
\item{x}{univariate observations from the finite population}

\item{wgts}{sampling weights}

\item{initial}{an initial estimate for the shape and scale parameters. If missing,
use the method of moments.}

\item{integration_subdivisions}{number of partitions to divide the domain of \eqn{\hat f()}
for Gauss-Kronrod quadrature.}

\item{bw}{bandwidth for the HT-adjusted KDE. By default,
use Scott's rule (see \code{\link[stats:bandwidth]{stats::bw.nrd()}}).}

\item{optim_method, optim_control}{method and control options passed on to \code{\link[stats:optim]{stats::optim()}}.}
}
\description{
Compute the MHDE for the shape \eqn{k} and scale \eqn{\lambda} of the Weibull model.
}
\seealso{
Other Minimum Hellinger Distance Estimator: 
\code{\link{mhd_gamma}()},
\code{\link{mhd_lognorm}()}
}
\concept{Minimum Hellinger Distance Estimator}
