% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhde_gamma.R
\name{mle_gamma}
\alias{mle_gamma}
\title{Maximum Likelihood Estimator for the Gamma Model}
\usage{
mle_gamma(x, wgts = NULL)
}
\arguments{
\item{x}{univariate observations from the finite population}

\item{wgts}{sampling weights}
}
\value{
a list with `estimates`, the covariance `cov` and the density function.
}
\description{
Maximum Likelihood Estimator for the Gamma Model
}
\keyword{internal}
