% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhde_lognormal.R
\name{mle_lognormal}
\alias{mle_lognormal}
\title{Maximum Likelihood Estimator for the Lognormal Model}
\usage{
mle_lognormal(x, wgts = NULL)
}
\arguments{
\item{x}{univariate observations from the finite population}

\item{wgts}{sampling weights}
}
\value{
a list with `estimates`, the covariance `cov` and the density function.
}
\description{
Maximum Likelihood Estimator for the Lognormal Model
}
\keyword{internal}
