% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhde_weibull.R
\name{mom_weibull}
\alias{mom_weibull}
\title{Method of Moments Estimator for the Weibull Model}
\usage{
mom_weibull(x, wgts, shape_range = c(0.02, 100), cov = FALSE)
}
\arguments{
\item{x}{univariate observations from the finite population}

\item{wgts}{sampling weights}
}
\value{
a list with \code{estimates}, the covariance \code{cov} and the density function.
}
\description{
Method of Moments Estimator for the Weibull Model
}
\keyword{internal}
