% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss-kronrod-basis.R
\name{partition_domain}
\alias{partition_domain}
\title{Partition the Domain Into Subdivisions}
\usage{
partition_domain(x, bw, subdivisions, range)
}
\arguments{
\item{x}{numeric vector}

\item{bw}{bandwidth of the bounded kernel function}

\item{subdivisions}{desired number of subdivisions}

\item{range}{the allowable range of `x`}
}
\value{
a list of intervals
}
\description{
This partitions the domain spanned by `x` into a fixed number of partitions such that the
KDE \eqn{\hat{f}(x) > 0} within each of those partitions and \eqn{\hat{f}(x) = 0} outside
those partitions.
The algorithm tries to have all partitions roughly the same length, but the
apportionment is greedy and hence likely not the best possible option.
}
\keyword{internal}
