% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum-phi-div-estimator.R
\name{phidiv_gauss_quadrature}
\alias{phidiv_gauss_quadrature}
\title{Functor to Create Fixed-grid Gaussian Quadrature For General Phi Divergences}
\usage{
phidiv_gauss_quadrature(
  x,
  wgts,
  bandwidth,
  n_subdivisions = 256,
  poly_order = 20,
  divergence,
  family,
  kernel
)
}
\value{
a list of two functions:
\code{divergence_int} computes the above integral (1), and
\code{sandwich_cov_A} which computes the integral (2)
}
\description{
This creates two function.
The function \code{divergence_int} to computes integral (1):
\deqn{\int g_\theta(x) \phi(\hat{f}(x) / g_\theta(x)) \;\mathrm{d}x,}
for model density \eqn{g_\theta()} and convex function \eqn{\phi(x)}.
The function is set up such that \eqn{\hat f} is evaluated a fixed number of times in each
subdivision, and does not need to be re-evaluated for different \eqn{g_\theta()}.
}
\details{
The function \code{sandwich_cov_A} to computes the integrals (2):
\deqn{\int g_\theta(x) \left[
   w_1(\hat{f}(x) / g_\theta(x)) H_\theta(x) +
   w_2(\hat{f}(x) / g_\theta(x)) s_\theta(x) s_\theta(x)^\intercal \right] \;\mathrm{d}x,
 }
for a model family with density \eqn{g_\theta()}, score function \eqn{s_\theta()}
and Hessian function \eqn{H_\theta()}.
The functions \eqn{w_j(x)} are associated with the phi divergence and defined as
\eqn{w_1(x) = \phi(x) - x \phi'(x)} and \eqn{w_2(x) = x^2 \phi''(x)}.
The function is set up such that \eqn{\hat f} is evaluated a fixed number of times in each
subdivision, and does not need to be re-evaluated for different members of the same model
family.
}
\keyword{internal}
