% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-divergence.R
\name{PhiDivergence}
\alias{PhiDivergence}
\title{PhiDivergence Class}
\description{
A class to encapsulate different phi divergences.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{full name of the divergence measure}

\item{\code{phi}}{divergence generator \eqn{\phi()}}

\item{\code{w1}}{function to evaluate \eqn{\phi(x) - x \phi'(x)}}

\item{\code{w2}}{function to evaluate \eqn{x^2 \phi''(x)}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PhiDivergence-new}{\code{PhiDivergence$new()}}
\item \href{#method-PhiDivergence-clone}{\code{PhiDivergence$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhiDivergence-new"></a>}}
\if{latex}{\out{\hypertarget{method-PhiDivergence-new}{}}}
\subsection{Method \code{new()}}{
Define a new phi divergence.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhiDivergence$new(name, phi, w1, w2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of phi divergence.}

\item{\code{phi}}{the generator function.
The generator must be convex function \eqn{\phi\colon [0, \infty) \to (-\infty, \infty]}
and satisfy \eqn{\phi(1) = 1}, \eqn{\lim_{t \to 0^+} \phi(t) = \phi(0)}.}

\item{\code{w1}}{function to evaluate \eqn{\phi(x) - x \phi'(x)}}

\item{\code{w2}}{function to evaluate \eqn{x^2 \phi''(x)}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhiDivergence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PhiDivergence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhiDivergence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
