% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum-phi-div-estimator.R
\name{survey_mpde}
\alias{survey_mpde}
\title{Minimum Phi Distance Estimator for the General Continuous Models with Complex
Survey Designs}
\usage{
survey_mpde(
  x,
  design,
  initial,
  family,
  divergence = "HD",
  na.rm = FALSE,
  integration_subdivisions = 256,
  bw,
  kernel = c("epanechnikov", "triangular", "rectangular", "biweight"),
  optim_method = "Nelder-Mead",
  optim_control = list(),
  mismatch_penalty = 10,
  singularity_penalty = mismatch_penalty
)
}
\arguments{
\item{x}{a formula, symbol, expression, vector, or matrix specifying the observations.
The objects are first looked up in the provided \code{design}, then in the calling environment.}

\item{design}{the survey design created by \code{\link[survey:svydesign]{survey::svydesign()}} and friends.}

\item{initial}{an initial estimate for the parameters.
If the vector has names, those will be used for the returned estimates.}

\item{family}{either the name of a known model family, or a
model family as returned by \code{\link[=model_family]{model_family()}}.}

\item{divergence}{either the name of a known phi divergence, or a
phi divergence as returned by \code{\link[=phi_divergence]{phi_divergence()}}.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values
should be omitted.}

\item{integration_subdivisions}{number of partitions to divide the domain of \eqn{\hat f()}
for Gauss-Kronrod quadrature.}

\item{bw}{bandwidth for the HT-adjusted KDE. Uses \code{\link[stats:bandwidth]{stats::bw.nrd()}} as default.}

\item{kernel}{Name of the kernel used in the HT-adjusted KDE.}

\item{optim_method, optim_control}{method and control options passed on to \code{\link[stats:optim]{stats::optim()}}.}

\item{mismatch_penalty, singularity_penalty}{penalties for mismatches between the scale
of the model density and the KDE, and for singularities of the KDE w.r.t. the mode
density. See details.}
}
\description{
Compute the survey MPDE for the parameters of a user-specified continuous model.
}
\details{
\subsection{Singularities and mismatches}{

When minimizing the divergence between the KDE and the model density, 2 issues can arise:
\enumerate{
\item The integration grid is designed such that the KDE integrates to 1. If a parameter
combination for the model density yields a density that is concentrated in a substantially
narrower region, the model density would not integrate to 1 with this integration grid.
Since this model density is likely a poor fit to the data anyways, the optimization
penalizes these densities by \code{mismatch_penalty * (1 - f_theta_int)^2},
where \code{f_theta_int} is the evaluated integral
of the model density. Typically, a mismatch penalty several times the maximum value of
the divergence works well.
\item For divergences with infinite gradient as \eqn{\delta(x) = \hat f(x) / g_\theta(x) \to \infty},
singularities will result in the divergence becoming infinite, too.
To escape these parameter regions of poor fit, the optimization can replace the infinite value
with \code{singularity_penalty * (1 + family$inverse_scale(params))}.
}

Note that the returned value for the minimum \eqn{\phi}-divergence does not use these
adjustments and returns the actual divergence between the KDE and the fitted model.
}
}
\seealso{
Other Minimum Phi-Divergence Estimator: 
\code{\link{survey_regression_mpde}()}
}
\concept{Minimum Phi-Divergence Estimator}
