% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpde_regression.R
\name{survey_regression_mpde}
\alias{survey_regression_mpde}
\title{Minimum Phi Distance Estimator for the Linear Regression Models with Complex
Survey Designs}
\usage{
survey_regression_mpde(
  x,
  design,
  initial,
  family,
  link,
  divergence = "HD",
  na.rm = FALSE,
  integration_subdivisions = 256,
  bw = bw.nrd,
  kernel = c("epanechnikov", "triangular", "rectangular", "biweight"),
  optim_method = "Nelder-Mead",
  optim_control = list()
)
}
\arguments{
\item{x}{a formula or matrix specifying the linear model.
The objects are first looked up in the provided \code{design}, then in the calling environment.}

\item{design}{the survey design created by \code{\link[survey:svydesign]{survey::svydesign()}} and friends.}

\item{initial}{a list with two components:
\code{coef} with the initial estimates for the regression coefficients and
\code{nuisance} with the initial values for the nuisance components.
The family's \code{nuisance_name} has information about the nuisance parameters for the
respective model family.
The initial values must be on the scale of the link function!}

\item{family}{either the name of a known model family, or a
model family as returned by \code{\link[=model_family]{model_family()}}.}

\item{link}{the link functions obtained with \code{\link[=link]{link()}} for the mean and nuisance components
of the linear regression model.
If missing, uses the default link for the family \code{family$default_link}.}

\item{divergence}{either the name of a known phi divergence, or a
phi divergence as returned by \code{\link[=phi_divergence]{phi_divergence()}}.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values
should be omitted.}

\item{integration_subdivisions}{number of partitions to divide the domain of \eqn{\hat f()}
for Gaussian quadrature.}

\item{bw}{bandwidth for the HT-adjusted KDE. Either a scalar bandwidth (not advised)
used for all groups, or (the name of) a function which will be used to cmpute the
bandwidth separately in each group. Uses \code{\link[stats:bandwidth]{stats::bw.nrd()}} as default.}

\item{kernel}{Name of the kernel used in the HT-adjusted KDE.}

\item{optim_method, optim_control}{method and control options passed on to \code{\link[stats:optim]{stats::optim()}}.}
}
\description{
Compute the survey MPDE for the mean regression parameters under a user-specified continuous
model for the conditional distribution \eqn{Y \mid X = x} for \strong{discrete} \eqn{X}.
}
\seealso{
Other Minimum Phi-Divergence Estimator: 
\code{\link{survey_mpde}()}
}
\concept{Minimum Phi-Divergence Estimator}
