% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{vcov}
\alias{vcov}
\alias{vcov.survey_mhde}
\title{Calculate Variance-Covariance Matrix for Minimum Divergence Estimators}
\usage{
\method{vcov}{survey_mhde}(object, type = c("sandwich", "model"), n = c("score", "kish"), ...)
}
\arguments{
\item{object}{a MDE fitted to survey data.}

\item{type}{which type of covariance matrix to compute.}

\item{n}{either how to estimate the effective sample size, or (if numeric) the
sample size.}

\item{...}{currently unused.}
}
\description{
Calculate Variance-Covariance Matrix for Minimum Divergence Estimators
}
