% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhde_gamma.R
\name{mhd_gamma}
\alias{mhd_gamma}
\title{Minimum Hellinger Distance Estimator for the Gamma Model}
\usage{
mhd_gamma(
  x,
  wgts = NULL,
  initial,
  cov_type = c("sandwich", "model"),
  integration_subdivisions = 256,
  bw,
  optim_method = "Nelder-Mead",
  optim_control = list()
)
}
\arguments{
\item{x}{univariate observations from the finite population}

\item{wgts}{sampling weights}

\item{initial}{an initial estimate for the shape and scale parameters. If missing,
use the MLE.}

\item{cov_type}{whether to use the sandwich estimator for the covariance matrix
the inverse Fisher information under the model.}

\item{integration_subdivisions}{number of partitions to divide the domain of \eqn{\hat f()}
for Gauss-Kronrod quadrature.}

\item{bw}{bandwidth for the HT-adjusted KDE. By default,
use Scott's rule (see \code{\link[stats:bandwidth]{stats::bw.nrd()}}).}

\item{optim_method, optim_control}{method and control options passed on to \code{\link[stats:optim]{stats::optim()}}.}
}
\description{
Compute the MHDE for the shape \eqn{\alpha} and scale \eqn{\theta} of the Gamma model.
}
\seealso{
\code{\link[stats:GammaDist]{stats::dgamma()}} for the parametrization by shape and scale.

Other Minimum Hellinger Distance Estimator: 
\code{\link{mhd_lognorm}()},
\code{\link{mhd_weibull}()}
}
\concept{Minimum Hellinger Distance Estimator}
