% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhde_lognormal.R
\name{mhd_lognorm}
\alias{mhd_lognorm}
\title{Minimum Hellinger Distance Estimator for the Lognormal Model}
\usage{
mhd_lognorm(
  x,
  wgts = NULL,
  initial,
  log_transform = FALSE,
  optim_method = "Nelder-Mead",
  optim_control = list()
)
}
\arguments{
\item{x}{univariate observations from the finite population}

\item{initial}{an initial estimate for the shape and scale parameters. If missing,
use the MLE.}

\item{log_transform}{whether to log-transform the data before estimation or not.}

\item{optim_method, optim_control}{method and control options passed on to [stats::optim()].}

\item{wgt}{sampling weights}
}
\description{
Minimum Hellinger Distance Estimator for the Lognormal Model
}
\seealso{
Other Minimum Hellinger Distance Estimator: 
\code{\link{mhd_gamma}()}
}
\concept{Minimum Hellinger Distance Estimator}
