% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhde_lognormal.R
\name{mhd_lognorm}
\alias{mhd_lognorm}
\title{Minimum Hellinger Distance Estimator for the Lognormal Model}
\usage{
mhd_lognorm(
  x,
  design,
  initial,
  sandwich_cov = TRUE,
  na.rm = FALSE,
  integration_subdivisions = 256,
  bw,
  kernel = c("epanechnikov", "triangular", "rectangular", "biweight"),
  optim_method = "Nelder-Mead",
  optim_control = list()
)
}
\arguments{
\item{x}{a formula, symbol, expression, vector, or matrix specifying the observations.
The objects are first looked up in the provided \code{design}, then in the calling environment.}

\item{design}{the survey design created by \code{\link[survey:svydesign]{survey::svydesign()}} and friends.}

\item{initial}{an initial estimate for the shape and scale parameters. If missing,
use the MLE.}

\item{sandwich_cov}{if \code{TRUE}, compute the sandwich estimator for the covariance matrix.
If \code{FALSE}, uses the inverse of the Fisher information under independence and the
effective number of observations estimated from the scores.
The only reason for setting this to \code{FALSE} is to increase computation speed when the
covariance matrix is not needed.
Otherwise, use \code{\link[=vcov]{vcov()}} to get the desired estimator of the covariance matrix.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values
should be omitted.}

\item{integration_subdivisions}{number of partitions to divide the domain of \eqn{\hat f()}
for Gauss-Kronrod quadrature.}

\item{bw}{bandwidth for the HT-adjusted KDE. Uses \code{\link[stats:bandwidth]{stats::bw.nrd()}} as default.}

\item{kernel}{Name of the kernel used in the HT-adjusted KDE.}

\item{optim_method, optim_control}{method and control options passed on to \code{\link[stats:optim]{stats::optim()}}.}
}
\description{
Compute the MHDE for the parameters \eqn{\mu,\sigma} of the Lognormal model.
}
\seealso{
\code{\link[stats:Lognormal]{stats::dlnorm()}} for the parametrization by \code{meanlog} \eqn{\mu} and
\code{sdlog} \eqn{\sigma}.

Other Minimum Hellinger Distance Estimator: 
\code{\link{mhd_gamma}()},
\code{\link{mhd_norm}()},
\code{\link{mhd_weibull}()},
\code{\link{survey_mhde}()}
}
\concept{Minimum Hellinger Distance Estimator}
