% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_family.R
\name{model_family}
\alias{model_family}
\title{Access Model Families}
\usage{
model_family(name, ...)
}
\arguments{
\item{name}{name of the model family.
If missing, list all currently known model families.}

\item{...}{passed on to initialize a new \link{ModelFamily}.}

\item{.register}{if defining a new model family, whether to register it in the
list of known model families.}
}
\value{
an object of type \link{ModelFamily}.
}
\description{
Get existing model families or define a new model family.
}
