% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi-divergence.R
\name{phi_divergence}
\alias{phi_divergence}
\alias{alpha_divergence}
\alias{power_divergence}
\title{Get or Define a Phi Divergence}
\usage{
phi_divergence(name, phi, w1, w2)

alpha_divergence(alpha)

power_divergence(lambda)
}
\arguments{
\item{name}{name of the phi divergence.
If missing, list all currently known divergences.}

\item{phi, w1, w2}{the generator of the phi divergence and function of its derivatives.
Passed on to initialize a new \link{PhiDivergence}.}

\item{alpha}{value for the \eqn{\alpha} parameter.}

\item{lambda}{value for the \eqn{\lambda} power parameter.}
}
\value{
an object of type \link{PhiDivergence}.
}
\description{
A phi divergence is defined through its generator \eqn{\phi} as
\deqn{D_\phi(p \| q) := \int q(x) \phi(p(x) / q(x)) \mathrm{d} x.}

The convenience function \code{alpha_divergence()} defines a specific
\eqn{\alpha} divergence with generator
\deqn{f(x) = \frac{x^\alpha - \alpha x - (1 - \alpha)}{\alpha (\alpha - 1)}.}

The convenience function \code{power_divergence()} defines a specific
power divergence with generator
\deqn{f(x) = \frac{1}{\lambda (\lambda + 1)} (x^{\lambda + 1} - 1).}
}
