% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum-phi-div-estimator.R
\name{survey_mpde}
\alias{survey_mpde}
\title{Minimum Phi Distance Estimator for the General Continuous Models with Complex
Survey Designs}
\usage{
survey_mpde(
  x,
  design,
  initial,
  family,
  divergence = "HD",
  na.rm = FALSE,
  integration_subdivisions = 256,
  bw,
  kernel = c("epanechnikov", "triangular", "rectangular", "biweight"),
  optim_method = "Nelder-Mead",
  optim_control = list()
)
}
\arguments{
\item{x}{a formula, symbol, expression, vector, or matrix specifying the observations.
The objects are first looked up in the provided \code{design}, then in the calling environment.}

\item{design}{the survey design created by \code{\link[survey:svydesign]{survey::svydesign()}} and friends.}

\item{initial}{an initial estimate for the parameters.
If the vector has names, those will be used for the returned estimates.}

\item{family}{either the name of a known model family, or a
model family as returned by \code{\link[=model_family]{model_family()}}.}

\item{divergence}{either the name of a known phi divergence, or a
phi divergence as returned by \code{\link[=phi_divergence]{phi_divergence()}}.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values
should be omitted.}

\item{integration_subdivisions}{number of partitions to divide the domain of \eqn{\hat f()}
for Gauss-Kronrod quadrature.}

\item{bw}{bandwidth for the HT-adjusted KDE. Uses \code{\link[stats:bandwidth]{stats::bw.nrd()}} as default.}

\item{kernel}{Name of the kernel used in the HT-adjusted KDE.}

\item{optim_method, optim_control}{method and control options passed on to \code{\link[stats:optim]{stats::optim()}}.}
}
\description{
Compute the survey MPDE for the parameters of a user-specified continuous model.
}
\concept{Minimum Phi-Divergence Estimator}
