% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhde_gamma.R, R/mhde_lognormal.R,
%   R/mhde_normal.R, R/mhde_weibull.R, R/vcov-internal.R
\name{vcov.survey_mde_gamma}
\alias{vcov.survey_mde_gamma}
\alias{vcov.survey_mde_lognorm}
\alias{vcov.survey_mde_norm}
\alias{vcov.survey_mde_weibull}
\alias{vcov}
\title{Calculate Variance-Covariance Matrix for Minimum Divergence Estimators}
\usage{
\method{vcov}{survey_mde_gamma}(object, type = c("sandwich", "model"), n = c("score", "kish"), ...)

\method{vcov}{survey_mde_lognorm}(object, type = c("sandwich", "model"), n = c("score", "kish"), ...)

\method{vcov}{survey_mde_norm}(object, type = c("sandwich", "model"), n = c("score", "kish"), ...)

\method{vcov}{survey_mde_weibull}(object, type = c("sandwich", "model"), n = c("score", "kish"), ...)
}
\arguments{
\item{object}{a MDE fitted to survey data.}

\item{type}{which type of covariance matrix to compute.}

\item{n}{either how to estimate the effective sample size, or (if numeric) the
sample size.}

\item{...}{currently unused.}
}
\description{
Calculate Variance-Covariance Matrix for Minimum Divergence Estimators
}
